+++ src/search/test_search.h
@@ -31,6 +31,7 @@ class CTestSearch {
     static bool test_early_exit();
     static bool test_anti_repetition();
     static bool test_scores();
+    static bool test_non_greedy_mate_scores();
     static bool test_positions();
 };