+++ src/universal_chess_interface/uci_protocol.h
@@ -9,6 +9,8 @@
 #include "../technical_functions/string_functions.h"
 #include "../technical_functions/string_tokenizer.h"
 
+constexpr int uci_first_movenumber = 1;
+
 class CUciProtocol {
   public:
     CUciProtocol();